Fix issue #15 - Same Image in Multiple Image Questions
    Validate was not implemented, returning always true. As a
consequence, not calling validate() method in the ContainerUI
focuslost() and not calling commitValue(), not writing the value of the
answer.

    Instead of using imageArray from OpenRosaCameraManager as a buffer,
we now write the image's string value to the element and clear
imageArray afterwards. Also, reimplemented delete to not only clear the
imageArray but also write an empty string to the element.